requirements.txt
requests
python-dotenv
python-dateutil
beautifulsoup4
lxml
